+++ src/main/java/com/goatshriek/rubydragon/ruby/RubyGhidraInterpreter.java
@@ -168,6 +168,18 @@ public class RubyGhidraInterpreter extends ScriptableGhidraInterpreter {
 	}
 
 	/**
+	 * The name for the current function variable.
+	 *
+	 * @return The name for the current function variable.
+	 *
+	 * @since 4.0.0
+	 */
+	@Override
+	public String getCurrentFunctionName() {
+		return "$current_function";
+	}
+
+	/**
 	 * The name for the current highlight variable.
 	 *
 	 * @return The name for the current highlight variable.